ref: BoardSurfaceのリファクタリング
