+++ app/planning-poker/[rid]/components/utils/getKey.ts
@@ -1,0 +1,3 @@
+export const getKeys = <T extends {[key: string]: unknown}>(obj: T): (keyof T)[] => {
+  return Object.keys(obj)
+}
